move protocol logic to zkchannels module in binary
This is to make #69 more possible. Moves all the logic except the basic
match on the CLI into a module called zkchannels

Fixes all the imports, etc.
