+++ src/customer.rs
@@ -12,6 +12,7 @@ pub use crate::database::customer as database;
 pub use crate::defaults::customer as defaults;
 pub use crate::transport::client::{self as client, Chan, Client};
 pub use crate::transport::server::{self as server, Server};
+pub use crate::zkchannels::customer as zkchannels;
 
 #[derive(Debug, Clone, sqlx::Type, Serialize, Deserialize)]
 #[sqlx(transparent)]
