+++ src/merchant.rs
@@ -3,3 +3,4 @@ pub use crate::config::{merchant as config, merchant::Config};
 pub use crate::database::merchant as database;
 pub use crate::defaults::merchant as defaults;
 pub use crate::transport::server::{self as server, Chan, Server};
+pub use crate::zkchannels::merchant as zkchannels;
