+++ src/zkchannels/merchant/close.rs
@@ -3,7 +3,7 @@ use {anyhow::Context, async_trait::async_trait};
 
 use super::{database, load_tezos_client, Command};
 
-use zeekoe::{
+use crate::{
     abort,
     merchant::{
         cli,
