+++ src/zkchannels/merchant/manage.rs
@@ -1,12 +1,12 @@
 use super::{database, Command};
-use serde_json::json;
-use zeekoe::{
+use crate::{
     amount::Amount,
     merchant::{
         cli::{List, Show},
         Config,
     },
 };
+use serde_json::json;
 use {
     anyhow::Context,
     async_trait::async_trait,
